Moved organ model source files to the model subdirectories
