+++ src/grandorgue/CMakeLists.txt
@@ -113,6 +113,18 @@ midi/GOMidiPlayerContent.cpp
 midi/GOMidiSender.cpp
 midi/GOMidiReceiver.cpp
 midi/GOMidiRecorder.cpp
+model/GODummyPipe.cpp
+model/GOEnclosure.cpp
+model/GOManual.cpp
+model/GOModel.cpp
+model/GOPipe.cpp
+model/GORank.cpp
+model/GOReferencePipe.cpp
+model/GOSoundingPipe.cpp
+model/GOStop.cpp
+model/GOSwitch.cpp
+model/GOTremulant.cpp
+model/GOWindchest.cpp
 sound/ports/GOSoundJackPort.cpp
 sound/ports/GOSoundPort.cpp
 sound/ports/GOSoundPortFactory.cpp
@@ -148,9 +160,7 @@ GOBitmapCache.cpp
 GOCoupler.cpp
 GODC.cpp
 GODivisionalCoupler.cpp
-GODummyPipe.cpp
 GODrawStop.cpp
-GOEnclosure.cpp
 GOEvent.cpp
 GOFilename.cpp
 GOFont.cpp
@@ -158,23 +168,13 @@ GOKeyReceiver.cpp
 GOLoadThread.cpp
 GOLog.cpp
 GOLogWindow.cpp
-GOManual.cpp
 GOMetronome.cpp
 GOMainWindowData.cpp
-GOModel.cpp
-GOPipe.cpp
 GOPipeConfig.cpp
 GOPipeConfigNode.cpp
 GOPipeConfigTreeNode.cpp
 GOProperties.cpp
-GORank.cpp
-GOReferencePipe.cpp
 GOReleaseAlignTable.cpp
-GOSoundingPipe.cpp
-GOStop.cpp
-GOTremulant.cpp
-GOWindchest.cpp
-GOSwitch.cpp
 GODefinitionFile.cpp
 GOFrame.cpp
 GODocument.cpp
