+++ src/grandorgue/GODefinitionFile.cpp
@@ -17,6 +17,7 @@
 #include "archive/GOArchive.h"
 #include "archive/GOArchiveFile.h"
 #include "archive/GOArchiveManager.h"
+#include "combinations/GODivisionalSetter.h"
 #include "combinations/GOSetter.h"
 #include "combinations/control/GOGeneralButtonControl.h"
 #include "config/GOConfig.h"
@@ -45,6 +46,13 @@
 #include "midi/GOMidiEvent.h"
 #include "midi/GOMidiPlayer.h"
 #include "midi/GOMidiRecorder.h"
+#include "model/GOEnclosure.h"
+#include "model/GOManual.h"
+#include "model/GORank.h"
+#include "model/GOSoundingPipe.h"
+#include "model/GOSwitch.h"
+#include "model/GOTremulant.h"
+#include "model/GOWindchest.h"
 #include "sound/GOSoundEngine.h"
 #include "temperaments/GOTemperament.h"
 
@@ -55,23 +63,15 @@
 #include "GOCoupler.h"
 #include "GODivisionalCoupler.h"
 #include "GODocument.h"
-#include "GOEnclosure.h"
 #include "GOEvent.h"
 #include "GOFile.h"
 #include "GOFilename.h"
 #include "GOHash.h"
 #include "GOLoadThread.h"
-#include "GOManual.h"
 #include "GOMetronome.h"
 #include "GOOrgan.h"
 #include "GOPath.h"
-#include "GORank.h"
 #include "GOReleaseAlignTable.h"
-#include "GOSoundingPipe.h"
-#include "GOSwitch.h"
-#include "GOTremulant.h"
-#include "GOWindchest.h"
-#include "combinations/GODivisionalSetter.h"
 
 GODefinitionFile::GODefinitionFile(GODocument *doc, GOConfig &settings)
   : m_doc(doc),
