+++ src/grandorgue/GODrawStop.cpp
@@ -9,10 +9,11 @@
 
 #include <wx/intl.h>
 
-#include "GODefinitionFile.h"
-#include "GOSwitch.h"
 #include "config/GOConfigReader.h"
 #include "config/GOConfigWriter.h"
+#include "model/GOSwitch.h"
+
+#include "GODefinitionFile.h"
 
 const struct IniFileEnumEntry GODrawstop::m_function_types[] = {
   {wxT("Input"), FUNCTION_INPUT},
