+++ src/grandorgue/gui/GOGUICouplerPanel.cpp
@@ -9,6 +9,8 @@
 
 #include <wx/intl.h>
 
+#include "model/GOManual.h"
+
 #include "GOCoupler.h"
 #include "GODefinitionFile.h"
 #include "GOGUIButton.h"
@@ -17,7 +19,6 @@
 #include "GOGUILayoutEngine.h"
 #include "GOGUIPanel.h"
 #include "GOGUISetterDisplayMetrics.h"
-#include "GOManual.h"
 
 GOGUICouplerPanel::GOGUICouplerPanel(GODefinitionFile *organfile)
   : m_organfile(organfile) {}
