+++ src/grandorgue/model/GOEnclosure.cpp
