+++ src/grandorgue/model/GOEnclosure.h
