+++ src/grandorgue/model/GOManual.cpp
