+++ src/grandorgue/model/GOPipe.h
