+++ src/grandorgue/model/GOStop.cpp
