+++ src/grandorgue/model/GOStop.h
