+++ src/grandorgue/model/GOSwitch.cpp
